requirements.txt
python-telegram-bot==7.0.1
configparser==3.5.0
SQLAlchemy==1.1.13
requests==2.18.3
alembic==0.9.4
psycopg2==2.7.3
